fix: address code review feedback on public-api session endpoints
- Use isValidKubernetesName() for session name validation instead of
  SanitizeForLog(); keep sanitized copy only for log.Printf calls
- Change NestedFloat64 to NestedInt64 for integer-backed fields
  (timeout, sdkRestartCount) in sessions_metrics.go
- Move t.Errorf out of HTTP handler goroutine in patch test
- Add label count assertion in transformSession test
- Change SessionResponse.Timeout from int to *int so timeout=0 is
  not swallowed by omitempty
- Add IMAGE_STATE_SYNC to imagePullPolicy override condition in deploy.sh
- Fix pre-existing ExtractServiceAccountFromAuth test to use X-Remote-User
  header (broken by f00505a refactor)

+++ components/backend/handlers/sessions_sub_test.go
@@ -97,7 +97,7 @@ var _ = Describe("Session Sub-Resource Handlers", Label(test_constants.LabelUnit
 
 				_ = unstructured.SetNestedField(session.Object, "Running", "status", "phase")
 				_ = unstructured.SetNestedField(session.Object, "2026-03-04T10:00:00Z", "status", "startTime")
-				_ = unstructured.SetNestedField(session.Object, float64(300), "spec", "timeout")
+				_ = unstructured.SetNestedField(session.Object, int64(300), "spec", "timeout")
 
 				_, err := k8sUtils.DynamicClient.Resource(sessionGVR).Namespace(testNamespace).Create(
 					ctx, session, v1.CreateOptions{})
